feat: エラーの定義を別ファイルに移動
