+++ GourmetApp/Model/ErrorModel.swift
@@ -1,0 +1,19 @@
+//
+//  ErrorModel.swift
+//  GourmetApp
+//
+//  Created by 藤井紗良 on 2024/02/08.
+//
+
+import Foundation
+
+enum APIError: Error {
+    case failCreateURL
+    case sessionError
+    case requestError(Error)
+    case decodeError
+}
+
+enum DataError: Error {
+    case imageError
+}
